Move DeclarativeRefresh to Document Update test expectations
