+++ tests/wpt/meta/html/browsers/browsing-the-web/navigating-across-documents/refresh/navigate.window.js.ini
@@ -2,6 +2,3 @@
   expected: TIMEOUT
   [When navigating the Refresh header needs to be followed]
     expected: TIMEOUT
-
-  [When there's both a Refresh header and <meta> the Refresh header wins]
-    expected: FAIL
