--- tests/wpt/meta/referrer-policy/generic/refresh-cross-origin.sub.html.ini
@@ -1,52 +1,0 @@
-[refresh-cross-origin.sub.html]
-  expected: TIMEOUT
-  [cross-origin header refresh with referrer policy "no-referrer" refreshes with the empty string as referrer]
-    expected: TIMEOUT
-
-  [cross-origin meta refresh with referrer policy "no-referrer-when-downgrade" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [cross-origin header refresh with referrer policy "no-referrer-when-downgrade" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [cross-origin meta refresh with referrer policy "origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin header refresh with referrer policy "origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin meta refresh with referrer policy "origin-when-cross-origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin header refresh with referrer policy "origin-when-cross-origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin meta refresh with referrer policy "same-origin" refreshes with the empty string as referrer]
-    expected: NOTRUN
-
-  [cross-origin header refresh with referrer policy "same-origin" refreshes with the empty string as referrer]
-    expected: NOTRUN
-
-  [cross-origin meta refresh with referrer policy "strict-origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin header refresh with referrer policy "strict-origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin meta refresh with referrer policy "strict-origin-when-cross-origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin header refresh with referrer policy "strict-origin-when-cross-origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin meta refresh with referrer policy "unsafe-url" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [cross-origin header refresh with referrer policy "unsafe-url" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [cross-origin meta refresh with referrer policy "" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [cross-origin header refresh with referrer policy "" refreshes with origin as referrer]
-    expected: NOTRUN
