--- tests/wpt/meta/referrer-policy/generic/refresh-same-url.html.ini
@@ -1,52 +1,0 @@
-[refresh-same-url.html]
-  expected: TIMEOUT
-  [same-URL header refresh with referrer policy "no-referrer" refreshes with the empty string as referrer]
-    expected: TIMEOUT
-
-  [same-URL meta refresh with referrer policy "no-referrer-when-downgrade" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL header refresh with referrer policy "no-referrer-when-downgrade" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL meta refresh with referrer policy "origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [same-URL header refresh with referrer policy "origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [same-URL meta refresh with referrer policy "origin-when-cross-origin" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL header refresh with referrer policy "origin-when-cross-origin" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL meta refresh with referrer policy "same-origin" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL header refresh with referrer policy "same-origin" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL meta refresh with referrer policy "strict-origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [same-URL header refresh with referrer policy "strict-origin" refreshes with origin as referrer]
-    expected: NOTRUN
-
-  [same-URL meta refresh with referrer policy "strict-origin-when-cross-origin" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL header refresh with referrer policy "strict-origin-when-cross-origin" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL meta refresh with referrer policy "unsafe-url" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL header refresh with referrer policy "unsafe-url" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL meta refresh with referrer policy "" refreshes with full url as referrer]
-    expected: NOTRUN
-
-  [same-URL header refresh with referrer policy "" refreshes with full url as referrer]
-    expected: NOTRUN
